Starting to patch Anti-Piracy.. Moved onto its own branch since this will be a bumpy ride...
So far this is a basic setup of editing the ROM at a single location where a value is stored, which is *NOP*'d out, which essentially makes the one variable in the LAB function useless. (I think..) Gonna have to continue to add the rest of the values and speedrun the game to see if I can get it to trigger.