Rework time again!
I decided to rework the main program again for my sanity, and just because its a new year and I wanted to cleanup the program and make it easier to use.
This push includes:
An updated .gitignore since newer Beef updates include a recovery/ folder, so I updated the .gitignore to include it.
A fresh new Program.bf that includes the updated usage.
And a backup of the Old Program.bf just to use as reference and something to fall back onto if something goes wrong.