The beginning of Hell has started.
This push includes the beginning of Prototype support. Going from Prototypes, to Pre and Post Release Demos, and all the way to Localisation Demos. There is a very slim chance of me being able to support every build, but I will attempt my best. So far this push only includes support for 2 builds, Spyro: Year of the Dragon - April 25th Prototype, and Spyro: Year of the Dragon - September 4th Prototype. The rest this push has is future proofing for all the builds for now, such as all the string names and updating the integer arrays to 37 to support all the versions. I also supported another type of file in the File Filter, .IMG files now!